修正

+++ ex00/megaphone.cpp
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*   Megaphone.cpp                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: skomatsu <skomatsu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/18 19:37:24 by skomatsu          #+#    #+#             */
-/*   Updated: 2025/09/18 20:45:11 by skomatsu         ###   ########.fr       */
+/*   Updated: 2025/12/15 19:43:12 by skomatsu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int main (int argc, char **argv)
         for (int i = 1; i < argc; i++)
         {
             std::string str(argv[i]);
-            for (int j = 0; j < str.length(); j++)
+            for (size_t j = 0; j < str.length(); j++)
                 std::cout << (char)std::toupper(str[j]);
         }
         std::cout << std::endl;
